objects/pickup: merge pickup collision in TRX
This merges pickup collision in TRX, along with animated interactions
for TR2. This is excluded for flare pickups currently.
