+++ src/libtrx/config/map.def
@@ -51,6 +51,7 @@ X_CFG_BOOL(gameplay.enable_target_change,            true)
 X_CFG_BOOL(gameplay.enable_timer_in_inventory,       TR_VERSION >= 2)
 X_CFG_BOOL(gameplay.enable_tr2_swimming,             true)
 X_CFG_BOOL(gameplay.enable_uw_roll,                  true)
+X_CFG_BOOL(gameplay.enable_walk_to_items,            false)
 X_CFG_BOOL(gameplay.fix_alligator_ai,                true)
 X_CFG_BOOL(gameplay.fix_bear_ai,                     true)
 X_CFG_BOOL(gameplay.fix_bridge_collision,            true)
