+++ src/libtrx/config/map_tr1.def
@@ -15,7 +15,6 @@ X_CFG_BOOL(gameplay.enable_total_stats,        true)
 X_CFG_BOOL(gameplay.enable_tr2_jumping,        true)
 X_CFG_BOOL(gameplay.enable_tr2_swim_cancel,    true)
 X_CFG_BOOL(gameplay.enable_wading,             true)
-X_CFG_BOOL(gameplay.enable_walk_to_items,      false)
 X_CFG_BOOL(gameplay.fix_shotgun_targeting,     true)
 X_CFG_BOOL(gameplay.restore_ps1_enemies,       false)
 X_CFG_BOOL(input.enable_buffering,             false)
