+++ src/libtrx/game/objects/general/pickup.c
@@ -2,11 +2,20 @@
 
 #include "config.h"
 #include "game/effects.h"
+#include "game/gun.h"
+#include "game/inventory.h"
 #include "game/lara.h"
+#include "game/lua.h"
+#include "game/overlay.h"
 #include "game/random.h"
+#include "game/rooms.h"
 #include "game/savegame.h"
 
 // clang-format off
+#define M_LF_PICKUP_ERASE       42
+#define M_LF_PICKUP_FLARE       58
+#define M_LF_PICKUP_FLARE_UW    20
+#define M_LF_PICKUP_UW          18
 #define M_AID_DIST_MIN          (STEP_L * 5)      // 1280
 #define M_AID_DIST_MAX          (WALL_L * 8)      // 8192
 #define M_AID_WAIT_MIN          (LOGIC_FPS * 2.5) // 75
@@ -46,6 +55,11 @@ static const OBJECT_BOUNDS m_PickUpBoundsUW = {
         .max = { .x = +45 * DEG_1, .y = +45 * DEG_1, .z = +45 * DEG_1, },
     },
 };
+
+static const XYZ_32 m_PickupPosition = { .x = 0, .y = 0, .z = -100 };
+static const XYZ_32 m_PickupPositionUW = { .x = 0, .y = -200, .z = -350 };
+
+extern void Stats_AddPickup(void);
 
 static void M_Initialise(int16_t item_num)
 {
@@ -141,19 +155,251 @@ static void M_Control(int16_t item_num)
     item->priv = (void *)(intptr_t)(int32_t)timer;
 }
 
+static void M_DoPickup(const int16_t item_num)
+{
+    ITEM *const item = Item_Get(item_num);
+    if (item->object_id == O_FLARE_ITEM) {
+        return;
+    }
+
+    Overlay_AddDisplayPickup(item->object_id);
+    Inv_AddPickup(item);
+#if TR_VERSION == 1
+    Stats_AddPickup();
+#endif
+    // Notify Lua pickup listeners
+    Lua_FireEvent(LUA_EVENT_PICKUP, item_num); // LUA uses 1-indexing
+
+    item->status = IS_INVISIBLE;
+    item->flags |= IF_KILLED;
+    Item_RemoveDrawn(item_num);
+    Item_RemoveActive(item_num);
+
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->interact_target.is_moving = false;
+}
+
+static void M_DoFlarePickup(const int16_t item_num)
+{
+    const ITEM *const item = Item_Get(item_num);
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->request_gun_type = LGT_FLARE;
+    lara->gun_type = LGT_FLARE;
+    Gun_InitialiseNewWeapon();
+    lara->gun_status = LGS_SPECIAL;
+#if TR_VERSION >= 2
+    lara->flare.age = ((int32_t)(intptr_t)item->data) & 0x7FFF;
+#endif
+    Item_Kill(item_num);
+    lara->interact_target.is_moving = false;
+}
+
+static void M_GetAllAtLaraPos(const ITEM *const item)
+{
+    int16_t pickup_num = Room_Get(item->room_num)->item_num;
+    while (pickup_num != NO_ITEM) {
+        ITEM *const check_item = Item_Get(pickup_num);
+        if (check_item->pos.x == item->pos.x && check_item->pos.z == item->pos.z
+            && Object_Get(check_item->object_id)->collision_func
+                == Pickup_Collision) {
+            M_DoPickup(pickup_num);
+        }
+        pickup_num = check_item->next_item;
+    }
+}
+
+static void M_DoControlled(const int16_t item_num, ITEM *const lara_item)
+{
+    ITEM *const item = Item_Get(item_num);
+    const XYZ_16 old_rot = item->rot;
+
+    item->rot.x = 0;
+    item->rot.y = lara_item->rot.y;
+    item->rot.z = 0;
+
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    if ((g_Input.action && lara->gun_status == LGS_ARMLESS
+         && !lara_item->gravity && lara_item->current_anim_state == LS(LS_STOP)
+         && !lara->interact_target.is_moving)
+        || (lara->interact_target.is_moving
+            && lara->interact_target.item_num == item_num)) {
+        const OBJECT *const obj = Object_Get(item->object_id);
+        if (Lara_TestPosition(item, obj->bounds_func())) {
+            const XYZ_32 pos = {
+                .x = m_PickupPosition.x,
+                .y = lara_item->pos.y - item->pos.y,
+                .z = m_PickupPosition.z,
+            };
+            if (Lara_MovePosition(item, &pos)) {
+                Item_SwitchToAnim(lara_item, LA(LA_PICKUP), 0);
+                lara_item->current_anim_state = LS(LS_PICKUP);
+                lara->head_rot.y = 0;
+                lara->head_rot.x = 0;
+                lara->torso_rot.y = 0;
+                lara->torso_rot.x = 0;
+                lara->interact_target.is_moving = false;
+                lara->gun_status = LGS_HANDS_BUSY;
+            }
+            lara->interact_target.item_num = item_num;
+        } else if (
+            lara->interact_target.is_moving
+            && lara->interact_target.item_num == item_num) {
+            lara->interact_target.is_moving = false;
+            lara->interact_target.item_num = NO_ITEM;
+            lara->gun_status = LGS_ARMLESS;
+        }
+
+        goto cleanup;
+    }
+
+    if (lara->interact_target.item_num != item_num) {
+        goto cleanup;
+    }
+
+    if (lara_item->current_anim_state == LS(LS_PICKUP)) {
+        if (Item_TestFrameEqual(lara_item, M_LF_PICKUP_ERASE)) {
+            M_GetAllAtLaraPos(item);
+        }
+        goto cleanup;
+    }
+
+cleanup:
+    item->rot = old_rot;
+}
+
+static void M_DoAboveWater(const int16_t item_num, ITEM *const lara_item)
+{
+    ITEM *const item = Item_Get(item_num);
+    if (g_Config.gameplay.enable_walk_to_items
+        && item->object_id != O_FLARE_ITEM) {
+        M_DoControlled(item_num, lara_item);
+        return;
+    }
+
+    const OBJECT *const obj = Object_Get(item->object_id);
+    const XYZ_16 old_rot = item->rot;
+
+    item->rot.x = 0;
+    item->rot.y = lara_item->rot.y;
+    item->rot.z = 0;
+
+    if (!Lara_TestPosition(item, obj->bounds_func())) {
+        goto cleanup;
+    }
+
+    if (lara_item->current_anim_state == LS(LS_PICKUP)) {
+        if (Item_TestFrameEqual(lara_item, M_LF_PICKUP_ERASE)) {
+            M_DoPickup(item_num);
+        }
+        goto cleanup;
+    }
+
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    if (lara_item->current_anim_state == LS(LS_FLARE_PICKUP)) {
+        if (Item_TestFrameEqual(lara_item, M_LF_PICKUP_FLARE)
+            && item->object_id == O_FLARE_ITEM && lara->gun_type != LGT_FLARE) {
+            M_DoFlarePickup(item_num);
+        }
+        goto cleanup;
+    }
+
+    if (g_Input.action && !lara_item->gravity
+        && lara_item->current_anim_state == LS(LS_STOP)
+        && lara->gun_status == LGS_ARMLESS
+        && (lara->gun_type != LGT_FLARE || item->object_id != O_FLARE_ITEM)) {
+        if (item->object_id == O_FLARE_ITEM) {
+            Lara_AnimateUntil(lara_item, LS(LS_FLARE_PICKUP));
+        } else {
+            Lara_AlignPosition(item, &m_PickupPosition);
+            Lara_AnimateUntil(lara_item, LS(LS_PICKUP));
+        }
+        lara_item->goal_anim_state = LS(LS_STOP);
+        lara->gun_status = LGS_HANDS_BUSY;
+        goto cleanup;
+    }
+
+cleanup:
+    item->rot = old_rot;
+}
+
+static void M_DoUnderwater(const int16_t item_num, ITEM *const lara_item)
+{
+    ITEM *const item = Item_Get(item_num);
+    const OBJECT *const obj = Object_Get(item->object_id);
+    const XYZ_16 old_rot = item->rot;
+
+    item->rot.x = -25 * DEG_1;
+    item->rot.y = lara_item->rot.y;
+    item->rot.z = 0;
+
+    if (!Lara_TestPosition(item, obj->bounds_func())) {
+        goto cleanup;
+    }
+
+    if (lara_item->current_anim_state == LS(LS_PICKUP)) {
+        if (Item_TestFrameEqual(lara_item, M_LF_PICKUP_UW)) {
+            M_DoPickup(item_num);
+        }
+        goto cleanup;
+    }
+
+    const LARA_INFO *const lara = Lara_GetLaraInfo();
+    if (lara_item->current_anim_state == LS(LS_FLARE_PICKUP)) {
+        if (Item_TestFrameEqual(lara_item, M_LF_PICKUP_FLARE_UW)
+            && item->object_id == O_FLARE_ITEM && lara->gun_type != LGT_FLARE) {
+            M_DoFlarePickup(item_num);
+            Lara_Flare_DrawMeshes();
+        }
+        goto cleanup;
+    }
+
+    if (g_Input.action && lara_item->current_anim_state == LS(LS_TREAD)
+        && lara->gun_status == LGS_ARMLESS
+        && (lara->gun_type != LGT_FLARE || item->object_id != O_FLARE_ITEM)) {
+        if (!Lara_MovePosition(item, &m_PickupPositionUW)) {
+            goto cleanup;
+        }
+
+        if (item->object_id == O_FLARE_ITEM) {
+            lara_item->fall_speed = 0;
+            Item_SwitchToAnim(lara_item, LA(LA_UNDERWATER_FLARE_PICKUP), 0);
+            lara_item->current_anim_state = LS(LS_FLARE_PICKUP);
+        } else {
+            if (g_Config.gameplay.fix_lara_pickup_embed) {
+                lara_item->fall_speed = 0;
+            }
+            Lara_AnimateUntil(lara_item, LS(LS_PICKUP));
+        }
+        lara_item->goal_anim_state = LS(LS_TREAD);
+        goto cleanup;
+    }
+
+cleanup:
+    item->rot = old_rot;
+}
+
+static void M_Setup(OBJECT *const obj)
+{
+    obj->activate_func = M_Activate;
+    obj->control_func = M_Control;
+    obj->collision_func = Pickup_Collision;
+    obj->bounds_func = Pickup_Bounds;
+    obj->draw_func = Object_DrawPickupItem;
+    obj->initialise_func = M_Initialise;
+    obj->handle_save_func = M_HandleSave;
+    obj->save_position = true;
+    obj->save_flags = true;
+}
+
 const OBJECT_BOUNDS *Pickup_Bounds(void)
 {
     const LARA_INFO *const lara = Lara_GetLaraInfo();
     if (lara->water_status == LWS_UNDERWATER
         || lara->water_status == LWS_CHEAT) {
         return &m_PickUpBoundsUW;
-    }
-#if TR_VERSION == 1
-    else if (g_Config.gameplay.enable_walk_to_items) {
+    } else if (g_Config.gameplay.enable_walk_to_items) {
         return &m_PickUpBoundsControlled;
-    }
-#endif
-    else {
+    } else {
         return &m_PickUpBounds;
     }
 }
@@ -169,17 +415,23 @@ bool Pickup_Trigger(const int16_t item_num)
     return true;
 }
 
-static void M_Setup(OBJECT *const obj)
+void Pickup_Collision(
+    const int16_t item_num, ITEM *const lara_item, COLL_INFO *const coll)
 {
-    obj->activate_func = M_Activate;
-    obj->control_func = M_Control;
-    obj->collision_func = Pickup_Collision;
-    obj->bounds_func = Pickup_Bounds;
-    obj->draw_func = Object_DrawPickupItem;
-    obj->initialise_func = M_Initialise;
-    obj->handle_save_func = M_HandleSave;
-    obj->save_position = true;
-    obj->save_flags = true;
+    const ITEM *const item = Item_Get(item_num);
+    if ((item->flags & IF_INVISIBLE) != 0) {
+        return;
+    }
+
+    const LARA_INFO *const lara = Lara_GetLaraInfo();
+    if (lara->water_status == LWS_ABOVE_WATER
+        || lara->water_status == LWS_WADE) {
+        M_DoAboveWater(item_num, lara_item);
+    } else if (
+        lara->water_status == LWS_UNDERWATER
+        || lara->water_status == LWS_CHEAT) {
+        M_DoUnderwater(item_num, lara_item);
+    }
 }
 
 REGISTER_OBJECT(O_EXPLOSIVE_ITEM, M_Setup)
