+++ src/tr2/meson.build
@@ -119,7 +119,6 @@ sources = [
   'game/objects/general/final_cutscene.c',
   'game/objects/general/flare_item.c',
   'game/objects/general/harpoon_bolt.c',
-  'game/objects/general/pickup.c',
   'game/objects/general/switch.c',
   'game/objects/general/zipline.c',
   'game/objects/setup.c',
